Adjusted Nether Spawn
Vanilla Nether can be creepy, but it’s ultimately quite dull. Nothing
but Pig Zombies as far as the eye can see. I’ve turned the spawn rates
for Pig Zombies and Ghasts through the FLOOR (previously 100 and 50, on
a scale of 1 to 10).

I also added Zombies and Skeletons to the Nether to make things more
interesting. Further adjustments may be needed.